refactor(input): move keyboard dispatch + focus/gate markers to src/input/, delete ozma_terminal::input
Move `TerminalInputBindings`/`ReservedChord` → `src/input/bindings.rs`,
`KeyboardFocused`/`KeyboardDisabled` → `src/input/focus.rs`, and
`dispatch_input`/`chord_matches`/`bevy_key_to_terminal_key` →
`src/input/keyboard.rs` with `KeyboardInputPlugin`. Move
`OzmaTerminalInputSet` definition to `ozma_terminal/src/lib.rs` (anchor
preserved), then delete `crates/ozma_terminal/src/input.rs`. Repoint all
nine `KeyboardFocused` importers and three `KeyboardDisabled` importers to
`crate::input::focus`. Register `KeyboardInputPlugin` in `src/main.rs`.
