+++ crates/ozma_terminal/src/lib.rs
@@ -4,27 +4,28 @@ mod action;
 mod clipboard;
 mod exit;
 mod hyperlink;
-mod input;
 mod layout;
 mod mouse;
 mod spawn;
 
 use crate::action::OzmaActionPlugin;
-use crate::input::OzmaInputPlugin;
 use crate::mouse::OzmaMousePlugin;
 use crate::spawn::on_add_inject_render;
 use crate::{exit::ExitPlugin, layout::LayoutPlugin};
 pub use action::PasteAction;
 use bevy::prelude::*;
 pub use clipboard::{Clipboard, build_paste_bytes};
-pub use input::{
-    KeyboardDisabled, KeyboardFocused, OzmaTerminalInputSet, ReservedChord, TerminalInputBindings,
-};
 pub use mouse::{MouseEffect, OzmaTerminalMouseSet, TerminalForwardInput, TerminalMouseEffects};
 pub use spawn::{
     OzmaSpawnOptions, OzmaTerminal, OzmaTerminalBundle, OzmaTerminalConfig, cells_for,
     resolve_shell,
 };
+
+/// System set containing the default terminal keyboard dispatcher. Hosts that
+/// maintain `KeyboardDisabled` should schedule their maintainer
+/// `.before(OzmaTerminalInputSet)`.
+#[derive(SystemSet, Debug, Clone, PartialEq, Eq, Hash)]
+pub struct OzmaTerminalInputSet;
 
 /// Bevy plugin that registers the Ozma VT terminal subsystems.
 ///
@@ -41,13 +42,7 @@ impl Plugin for OzmaTerminalPlugin {
         app.insert_resource(OzmaTerminalConfig {
             shell: self.config_shell.clone(),
         })
-        .add_plugins((
-            ExitPlugin,
-            LayoutPlugin,
-            OzmaActionPlugin,
-            OzmaInputPlugin,
-            OzmaMousePlugin,
-        ))
+        .add_plugins((ExitPlugin, LayoutPlugin, OzmaActionPlugin, OzmaMousePlugin))
         .add_observer(on_add_inject_render);
     }
 }
