+++ src/mode/default/input.rs
@@ -5,6 +5,7 @@
 //! are owned by `ozma_terminal`'s dispatcher and `PasteAction`.
 
 use crate::input::focus::MouseDisabled;
+use crate::input::focus::{KeyboardDisabled, KeyboardFocused};
 use crate::input::ime::{ImeCommit, ImeState};
 use crate::input::shortcuts::ResolvedShortcuts;
 use crate::input::{InputPhase, current_modifiers};
@@ -19,9 +20,7 @@ use bevy::prelude::*;
 use bevy::ui::{ComputedNode, UiGlobalTransform};
 use bevy::window::{PrimaryWindow, Window};
 use bevy_cef::prelude::FocusedWebview;
-use ozma_terminal::{
-    KeyboardDisabled, KeyboardFocused, OzmaTerminal, OzmaTerminalInputSet, OzmaTerminalMouseSet,
-};
+use ozma_terminal::{OzmaTerminal, OzmaTerminalInputSet, OzmaTerminalMouseSet};
 use ozma_tty_engine::{TerminalKey, TerminalKeyInput, TerminalModifiers};
 use ozma_tty_renderer::TerminalCellMetricsResource;
 use ozma_tty_renderer::prelude::TerminalOverlays;
