+++ src/mode/tmux/gate.rs
@@ -6,6 +6,7 @@
 
 use super::pane_hit::tmux_pane_at_phys;
 use super::rename_prompt::RenamePrompt;
+use crate::input::focus::KeyboardDisabled;
 use crate::input::focus::MouseDisabled;
 use crate::input::ime::ImeState;
 use crate::mode::AppMode;
@@ -15,7 +16,7 @@ use bevy::prelude::*;
 use bevy::ui::{ComputedNode, UiGlobalTransform};
 use bevy::window::{PrimaryWindow, Window};
 use bevy_cef::prelude::FocusedWebview;
-use ozma_terminal::{KeyboardDisabled, OzmaTerminalInputSet, OzmaTerminalMouseSet};
+use ozma_terminal::{OzmaTerminalInputSet, OzmaTerminalMouseSet};
 use ozma_tty_renderer::TerminalCellMetricsResource;
 use ozma_tty_renderer::prelude::TerminalOverlays;
 use ozma_webview::{NonInteractive, Webview, webview_hit_at};
